casi finalizado

+++ daa/src/app/components/dashboard/reservaciones/reservaciones.component.ts
@@ -2,43 +2,52 @@ import { Component } from '@angular/core';
 import { UploadComponent } from './upload/upload.component';
 import { MatDialog } from '@angular/material/dialog';
 import { ReservationService } from 'src/app/services/reservation.service';
+import Swal from 'sweetalert2';
 
 interface Reservation {
   _id: string;
+  name: string;
   date: string;
   numOfPeople: number;
   time: string;
   email: string;
 }
 
-
 @Component({
   selector: 'app-reservaciones',
   templateUrl: './reservaciones.component.html',
-  styleUrls: ['./reservaciones.component.css']
+  styleUrls: ['./reservaciones.component.css'],
 })
-
 export class ReservacionesComponent {
-
   colors: string[] = ['#F44336', '#FF9800', '#2196F3', '#4CAF50'];
   reservations: Reservation[] = [];
 
-  constructor(private reservationService: ReservationService, private dialog: MatDialog){}
+  constructor(
+    private reservationService: ReservationService,
+    private dialog: MatDialog
+  ) {}
 
   editReservation(reservation: Reservation) {
     const dialogRef = this.dialog.open(UploadComponent, {
-      data: { reservation }
+      data: { reservation },
     });
     dialogRef.afterClosed().subscribe((result: Reservation) => {
       if (result) {
-        const updatedFields = { numOfPeople: result.numOfPeople, time: result.time };
-        console.log(reservation)
-        console.log(updatedFields)
-        this.reservationService.updateReservation(reservation._id, updatedFields)
+        const updatedFields = {
+          numOfPeople: result.numOfPeople,
+          time: result.time,
+        };
+        this.reservationService
+          .updateReservation(reservation._id, updatedFields)
           .subscribe(() => {
-            const index = this.reservations.findIndex(r => r._id === reservation._id);
+            const index = this.reservations.findIndex(
+              (r) => r._id === reservation._id
+            );
             if (index >= 0) {
-              const updatedReservation: Reservation = { ...reservation, ...updatedFields };
+              const updatedReservation: Reservation = {
+                ...reservation,
+                ...updatedFields,
+              };
               this.reservations[index] = updatedReservation;
             }
           });
@@ -47,15 +56,28 @@ export class ReservacionesComponent {
   }
 
   removeReservation(reservation: any): void {
-    const index = this.reservations.indexOf(reservation);
-    if (index >= 0) {
-      this.reservations.splice(index, 1);
-    }
-    this.reservationService
-      .eliminarReservacion(reservation)
-      .subscribe((res) => {
-        this.getAllReservations();
-      });
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Borrar!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const index = this.reservations.indexOf(reservation);
+        if (index >= 0) {
+          this.reservations.splice(index, 1);
+        }
+        this.reservationService
+          .eliminarReservacion(reservation)
+          .subscribe((res) => {
+            this.getAllReservations();
+          });
+        Swal.fire('Deleted!', 'Reservacion Borrada.', 'success');
+      }
+    });
   }
 
   getAllReservations(): void {
@@ -70,5 +92,4 @@ export class ReservacionesComponent {
   ngOnInit() {
     this.getAllReservations();
   }
-
 }
